requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1
dramatiq[redis, watch]
biom-format
typer[all]
dynaconf[all]
loguru
requests
aiohttp